refactor: reorganize app into src/app layout
