+++ src/app/components/create-workspace-modal.tsx
@@ -2,7 +2,7 @@ import { For, Show, createSignal } from "solid-js";
 
 import { CheckCircle2, FolderPlus, Loader2, X } from "lucide-solid";
 
-import Button from "./Button";
+import Button from "./button";
 
 export default function CreateWorkspaceModal(props: {
   open: boolean;
