+++ src/app/components/part-view.tsx
