+++ src/app/components/reset-modal.tsx
@@ -2,8 +2,8 @@ import { Match, Show, Switch } from "solid-js";
 
 import { X } from "lucide-solid";
 
-import Button from "./Button";
-import TextInput from "./TextInput";
+import Button from "./button";
+import TextInput from "./text-input";
 
 export type ResetModalProps = {
   open: boolean;
