+++ src/app/components/thinking-block.tsx
