+++ src/app/lib/opencode.ts
