+++ src/app/types.ts
@@ -1,6 +1,6 @@
 import type { Message, Part, PermissionRequest as ApiPermissionRequest, Provider, Session } from "@opencode-ai/sdk/v2/client";
-import type { createClient } from "../lib/opencode";
-import type { OpencodeConfigFile, WorkspaceInfo } from "../lib/tauri";
+import type { createClient } from "./lib/opencode";
+import type { OpencodeConfigFile, WorkspaceInfo } from "./lib/tauri";
 
 export type Client = ReturnType<typeof createClient>;
